query and client + server refactor
