+++ packages/cli/src/data/apollo-client.js
@@ -4,7 +4,7 @@ import { HttpLink } from 'apollo-link-http';
 
 const cache = new InMemoryCache();
 const link = new HttpLink({
-  uri: 'http://localhost:4000', // specific internal flag to prevent looping cache on request
+  uri: 'http://localhost:4000', // TODO specific internal flag to prevent looping cache on request
   cache: new InMemoryCache().restore(window.__APOLLO_STATE__) // eslint-disable-line no-underscore-dangle
 });
