+++ packages/cli/src/data/queries/hello.gql
@@ -1,0 +1,3 @@
+query HelloQuery {
+  hello
+}
